python_producer_image/requirements.txt
python-dateutil
kafka-python